Fix for AdminThemeUIKit Language Tabs
Uikit language tabs do not trigger same event as other admin themes, so JS was added to init the simple mde fields on specific event/selector to re-init the field. The re-init of the field is recommended instead of init on load/ready because of the refresh problem with code mirror; this ensures the field content is visible when the tab is clicked, without having to do refresh events or polling.